refactored the code of auth service
